requirements.txt
Django==4.1.2
gunicorn==20.1.0

psycopg2-binary==2.9.5
python-dotenv==0.21.0
python-decouple==3.6
requests==2.28.1

django-on-heroku==1.1.2
dj-database-url==1.0.0
whitenoise==6.2.0

djangorestframework==3.14.0
django-cors-headers==3.13.0